Stop archiving project and user identity.
Instead let's assume the project and user are setup already for us, and
we're importing a template into it. This should allow for more safety as
the new project may be different, attributes/annotations for how we
create users and such may have changed since archival, and this is more
generally useful for generating a template.